m6_modelos_no_supervisados_avanzados/AM6_anomalias_pacientes_cronicos/requirements.txt
numpy==2.3.1
pandas==2.3.0
scikit-learn==1.7.0
matplotlib==3.10.3
seaborn==0.13.2
umap-learn==0.5.6
hdbscan==0.8.38
kagglehub==0.2.5
